requirements.txt
# BOT_Trading v3.0 Dependencies
# Основные зависимости для мульти-трейдер системы

# Асинхронное программирование (проверено январь 2025)
aiohttp>=3.12.13  # ✅ Python 3.13 ready, stable production
aiofiles>=23.2.0  # ✅ Latest stable
asyncpg>=0.30.0   # ✅ PostgreSQL 9.5-17 support
websockets>=12.0  # ✅ WebSocket protocol compatibility

# Веб-сервер FastAPI (проверено январь 2025)
fastapi>=0.104.1        # ✅ Latest stable with Python 3.11+ support
uvicorn[standard]>=0.24.0  # ✅ ASGI server with auto-reload
python-multipart>=0.0.6   # ✅ Form data parsing
jinja2>=3.1.2           # ✅ Template engine (if needed)
python-jose[cryptography]>=3.3.0  # ✅ JWT tokens

# Криптобиржи (проверено январь 2025)
ccxt>=4.2.50           # ✅ 100+ exchanges, active updates
python-binance>=1.0.17 # ✅ Official Binance connector
bybit>=0.2.12          # ✅ Official Bybit SDK

# База данных (проверено январь 2025)
psycopg2-binary>=2.9.9  # ✅ Latest PostgreSQL driver
sqlalchemy>=2.0.25      # ✅ Async support, stable
alembic>=1.13.1         # ✅ Migration tool compatibility

# Машинное обучение (проверено январь 2025)
torch>=2.0.0         # ✅ Для PatchTST модели
xgboost>=2.0.3       # ⚠️ Breaking changes from 1.x, migration needed
pandas>=2.1.4        # ✅ Stable, broad compatibility
numpy>=1.24.4        # ✅ Python 3.11+ compatible
scikit-learn>=1.4.0  # ✅ Latest stable
joblib>=1.3.2        # ✅ Model serialization
scipy>=1.11.4        # ✅ Scientific computing

# Технические индикаторы
ta>=0.10.2          # ✅ Technical Analysis library (pure Python)
# ta-lib>=0.4.26  # Требует системную библиотеку TA-Lib
# talib>=0.4.26    # Альтернативное имя того же пакета

# Конфигурация (проверено январь 2025)
PyYAML>=6.0.1         # ✅ Security fixes included
pydantic>=2.5.3       # ⚠️ Breaking changes from 1.x, migration needed
python-dotenv>=1.0.0  # ✅ Environment management

# Логирование
loguru>=0.7.0
structlog>=23.1.0

# HTTP клиенты
requests>=2.31.0
httpx>=0.24.0

# WebSocket
websockets>=11.0.0
websocket-client>=1.5.0

# Криптография
cryptography>=40.0.0
python-jose>=3.3.0

# Мониторинг
prometheus-client>=0.17.0
sentry-sdk>=1.28.0,<2.0
urllib3>=1.26.11,<2.0  # Явно указываем версию для разрешения конфликтов

# Утилиты
click>=8.1.0
rich>=13.4.0
python-dateutil>=2.8.0
pytz>=2023.3
psutil>=5.9.0  # ✅ System monitoring utilities
python-telegram-bot>=20.0  # ✅ Telegram bot API
asyncio-throttle>=1.0.2  # ✅ Rate limiting для ML pipeline

# Тестирование
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Форматирование кода
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Документация
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Разработка
pre-commit>=3.3.0
bandit>=1.7.0
